microservice1_order_service/requirements.txt
Django
paho-mqtt
requests
